Corrected an error in the final scoring mechanism that was causing the losing player to have a score of 0.  Changed everything associated with the score to a long instead of an int.

+++ CS325 Project 3/src/oozeWars/Explosion.java
@@ -149,7 +149,7 @@ class Explosion extends Entity
 			for(Blob b : ow.getBlobs())
 			{
 				if(blobID ==  b.getBlobID())
-					b.addDamageDealt( (int)(tempDamage + .5) );
+					b.addDamageDealt( (long)(tempDamage + .5) );
 			}
 			
 			vx *= speed;
